requirements.txt
anyio==3.5.0
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
cheroot==8.6.0
CherryPy==18.6.1
colorama==0.4.4
decorator==5.1.1
Django==4.0.4
django-environ==0.8.1
django-mathfilters==1.0.0
dnspython==2.2.1
email-validator==1.1.3
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.14.7
httpx==0.22.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
iniconfig==1.1.1
ipaddress==1.0.23
jaraco.classes==3.2.1
jaraco.collections==3.5.1
jaraco.context==4.1.1
jaraco.functools==3.5.0
jaraco.text==3.7.0
loguru==0.5.3
more-itertools==8.12.0
packaging==21.3
phonenumbers==8.12.46
Pillow==9.1.0
pluggy==1.0.0
portend==3.1.0
priority==2.0.0
py==1.11.0
pydantic==1.8.2
pyparsing==3.0.8
pyQiwiP2P==2.0.0
pytest==6.2.5
python-dotenv==0.17.1
pytz==2022.1
requests==2.27.1
response-report==1.1.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.19.0
tempora==5.0.1
toml==0.10.2
typing==3.7.4.3
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
validators==0.18.2
win32-setctime==1.1.0
wsproto==1.1.0
zc.lockfile==2.0